Move topLevelResolveParser to a different file
